[ADD] 자주쓰는 타입 따로 분리 폴더 생성
