+++ src/components/review/ReviewMap.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function ReviewMap({ id }: { id: number }) {
+  return <div>ReviewMap</div>;
+}
+
+export default ReviewMap;
