+++ src/components/review/type/index.ts
@@ -1,0 +1,11 @@
+export type review_info = {
+  review_content: string | null;
+  review_createdat: string;
+  review_id: number | null;
+  toilet_clean_rate: number | null;
+  toilet_id: number | null;
+  toilet_loc_rate: number | null;
+  toilet_pop_rate: number | null;
+  user_id: string;
+  user_nickname: string | null;
+};
